Develop

+++ v2/2start_aosp_build.sh
@@ -39,23 +39,24 @@ touch $LOGPATH
 function w_log()
 {
 	s=`date "+%y%m%d_%H%M%S"`
-	echo  "$s: $1" |tee -a $LOGPATH
+	echo  "$s: $@" |tee -a $LOGPATH
 }
 
-w_log $@
-password=`grep password key.txt |awk -F " " '{print $2}' `
+w_log "$@"
 aliyun configure switch --profile default 1>/dev/null 2>&1
 w_log "step 1: create ecs instance"
+password=`grep password key.txt |awk -F " " '{print $2}' `
 #instance_id="i-0xi16uf7r4syux4z894e"
 if [ -z "$instance_id" ];then
 	instance_id=`aliyun ecs RunInstances --SpotStrategy SpotAsPriceGo --RegionId us-east-1 --InstanceType ecs.c9i.8xlarge --ImageId ubuntu_22_04_x64_20G_alibase_20230907.vhd \
-	--SecurityGroupId sg-0xi22iuffog8ylfh0or5  --VSwitchId vsw-0xi3vuydn6xui661xeiwu  --SystemDisk.Category cloud_essd  --SystemDisk.Size 350  --SystemDisk.PerformanceLevel PL1 \
-	--Password $password \
+	--SecurityGroupId sg-0xi22iuffog8ylfh0or5  --VSwitchId vsw-0xi3vuydn6xui661xeiwu  --SystemDisk.Category cloud_essd --SystemDisk.PerformanceLevel PL1 \
+       	--Password $password \
 	--InternetChargeType PayByTraffic --InternetMaxBandwidthOut 100 --Amount 1 --InstanceName openfde_aosp_make |jq -r .InstanceIdSets.InstanceIdSet[0] `
 fi
 
 if [[ -z "$instance_id" || "$instance_id" == "null" ]]; then
 	w_log "failed: ecs create failed"
+	sendEmail -xu 185457686@qq.com -xp guqbtpjnufzycbcb  -t 185457686@qq.com -s smtp.qq.com:587 -u "create ecs for aosp img making failed" -m "create ecs failed" -f 185457686@qq.com
 	exit 1
 fi
 sleep 30  # 等待实例启动
@@ -73,7 +74,17 @@ if [ "$instanceStatus" != '"Running"' ];then
 	w_log "step 2: after 180s ecs is  still not running"
 	exit 1
 fi
+disk_id=`aliyun ecs DescribeDisks --RegionId us-east-1 --DiskName aosp_source_disk_350G_$mode |jq -r .Disks.Disk[0].DiskId`
+if [[ -z "$disk_id" || "$disk_id" == "null" ]]; then
+	w_log "failed: disk id not found"
+	sendEmail -xu 185457686@qq.com -xp guqbtpjnufzycbcb  -t 185457686@qq.com -s smtp.qq.com:587 -u "disk id for img making not found" -m "disk id not found" -f 185457686@qq.com
+	w_log "delete instance $instance_id due to disk id not found"
+	aliyun ecs DeleteInstance --InstanceId $instance_id --Force true --RegionId us-east-1
+	exit 1
+fi
+aliyun ecs AttachDisk --InstanceId $instance_id --DiskId  $disk_id
 
+cd v2
 rm -rf build_aosp_shs.tgz
 w_log "step 3: tar build_aosp_shs.tgz"
 kid=`grep key-id key.txt |awk -F " " '{print $2}' `
@@ -84,7 +95,7 @@ sed -i "s/ACCESS_KEY_ID/$kid/g" wrapper_img.sh
 tar -zcvpf build_aosp_shs.tgz wrapper_img.sh make_imgs.sh .ssh/authorized_keys 1>/dev/null 2>&1
 
 
-shs=`cat /root/build_aosp_shs.tgz |base64`
+shs=`cat /root/v2/build_aosp_shs.tgz |base64`
 w_log "step 4: transferr build_aosp_shs.tgz"
 invoke_id=`aliyun ecs RunCommand  --Name "transfer_id" --Type "RunShellScript" --InstanceId.1 $instance_id  --RegionId us-east-1 \
 --CommandContent "echo '$shs' | base64 -d > /root/build_aosp_shs.tgz && tar -xf /root/build_aosp_shs.tgz -C /root/ && chmod +x /root/wrapper_img.sh" |jq -r '.InvokeId'`
@@ -124,16 +135,16 @@ if [ -z "$ip" ];then
 fi
 ssh-keygen -R $ip
 if [ "$mode" = "daily" ];then
-	ssh -o StrictHostKeyChecking=accept-new -o UserKnownHostsFile=/dev/null root@$ip  "setsid bash /root/wrapper_img.sh daily 1>/dev/null 2>&1 &"
+	ssh -o StrictHostKeyChecking=accept-new -o UserKnownHostsFile=/dev/null root@$ip  "setsid bash /root/wrapper_img.sh daily $2 1>/dev/null 2>&1 &"
 	if [ $? != 0 ];then
 		sleep 15
-		ssh -o StrictHostKeyChecking=accept-new -o UserKnownHostsFile=/dev/null root@$ip  "setsid bash /root/wrapper_img.sh daily 1>/dev/null 2>&1 &"
+		ssh -o StrictHostKeyChecking=accept-new -o UserKnownHostsFile=/dev/null root@$ip  "setsid bash /root/wrapper_img.sh daily $2 1>/dev/null 2>&1 &"
 	fi
 else
-	ssh -o StrictHostKeyChecking=accept-new -o UserKnownHostsFile=/dev/null root@$ip  "setsid bash /root/wrapper_img.sh version $2 $3 $4 $5 1>/dev/null 2>&1 &"
+	ssh -o StrictHostKeyChecking=accept-new -o UserKnownHostsFile=/dev/null root@$ip  "setsid bash /root/wrapper_img.sh version $2 $3 $4 $5 $6 1>/dev/null 2>&1 &"
 	if [ $? != 0 ];then
 		sleep 15
-		ssh -o StrictHostKeyChecking=accept-new -o UserKnownHostsFile=/dev/null root@$ip  "setsid bash /root/wrapper_img.sh version $2 $3 $4 $5 1>/dev/null 2>&1 &"
+		ssh -o StrictHostKeyChecking=accept-new -o UserKnownHostsFile=/dev/null root@$ip  "setsid bash /root/wrapper_img.sh version $2 $3 $4 $5 $6 1>/dev/null 2>&1 &"
 	fi
 fi
 if [ $? != 0 ];then
